+++ dbdaora/data_sources/fallback/datastore.py
@@ -1,5 +1,5 @@
 import dataclasses
-from typing import Any, Dict, Iterable, Optional, Sequence
+from typing import Any, Dict, Iterable, Optional
 
 from google.cloud.datastore import Client, Entity, Key
 
@@ -33,15 +33,6 @@ class DatastoreDataSource(FallbackDataSource[Key]):
 
     async def delete(self, key: Key) -> None:
         self.client.delete(key)
-
-    async def get_many(
-        self, keys: Iterable[Optional[Key]]
-    ) -> Sequence[Optional[Dict[str, Any]]]:
-        entities = self.client.get_multi([k for k in keys if k is not None])
-        goted_keys: Dict[Optional[Key], Entity] = {
-            entity.key: entity for entity in entities
-        }
-        return [goted_keys.get(key) for key in keys]
 
 
 def entity_asdict(entity: Entity) -> Dict[str, Any]: