+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_many.py
@@ -47,12 +47,93 @@ async def test_should_get_many_from_cache(
 
 
 @pytest.mark.asyncio
+async def test_should_get_many_from_fallback_when_not_found_on_memory(
+    fake_service, serialized_fake_entity, fake_entity, fake_entity2
+):
+    client = fake_service.repository.fallback_data_source.client
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake2'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:not-found:other_fake:fake'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:not-found:other_fake:fake2'
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake2'),
+        dataclasses.asdict(fake_entity2),
+    )
+
+    entities = await fake_service.get_many(
+        'fake', 'fake2', other_id='other_fake'
+    )
+
+    assert entities == [fake_entity, fake_entity2]
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake'
+    )
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake2'
+    )
+
+
+@pytest.mark.asyncio
+async def test_should_get_many_from_fallback_when_not_found_on_memory_with_fields(
+    fake_service, serialized_fake_entity, fake_entity, fake_entity2
+):
+    client = fake_service.repository.fallback_data_source.client
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake2'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:not-found:other_fake:fake'
+    )
+    await fake_service.repository.memory_data_source.delete(
+        'fake:not-found:other_fake:fake2'
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake2'),
+        dataclasses.asdict(fake_entity2),
+    )
+    fake_entity.number = None
+    fake_entity.boolean = False
+    fake_entity2.number = None
+    fake_entity2.boolean = False
+
+    entities = await fake_service.get_many(
+        'fake',
+        'fake2',
+        other_id='other_fake',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+    )
+
+    assert entities == [fake_entity, fake_entity2]
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake'
+    )
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake2'
+    )
+
+
+@pytest.mark.asyncio
 async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     fake_service, fake_entity, fake_entity2, mocker
 ):
-    fake_pipeline = mocker.MagicMock()
-    fake_service.repository.memory_data_source.pipeline = fake_pipeline
-    fake_pipeline.return_value.execute = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.hgetall = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     key = fake_service.repository.fallback_data_source.make_key(
@@ -85,10 +166,12 @@ async def test_should_get_many_with_fields(
     fake_entity2,
 ):
     await fake_service.repository.memory_data_source.hmset(
-        'fake:fake', *itertools.chain(*serialized_fake_entity.items())
+        'fake:other_fake:fake',
+        *itertools.chain(*serialized_fake_entity.items()),
     )
     await fake_service.repository.memory_data_source.hmset(
-        'fake:fake2', *itertools.chain(*serialized_fake_entity2.items())
+        'fake:other_fake:fake2',
+        *itertools.chain(*serialized_fake_entity2.items()),
     )
     entities = await fake_service.get_many(
         'fake',
@@ -132,9 +215,7 @@ async def test_should_get_many_from_cache_with_fields(
 async def test_should_get_many_from_fallback_after_open_circuit_breaker_with_fields(
     fake_service, fake_entity, fake_entity2, mocker
 ):
-    fake_pipeline = mocker.MagicMock()
-    fake_service.repository.memory_data_source.pipeline = fake_pipeline
-    fake_pipeline.return_value.execute = asynctest.CoroutineMock(
+    fake_service.repository.memory_data_source.hmget = asynctest.CoroutineMock(
         side_effect=RedisError
     )
     key = fake_service.repository.fallback_data_source.make_key('fake', 'fake')