+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_one.py
@@ -13,6 +13,7 @@ async def test_should_get_one(
     await fake_service.repository.memory_data_source.hmset(
         'fake:fake', *itertools.chain(*serialized_fake_entity.items())
     )
+
     entity = await fake_service.get_one('fake', other_id='other_fake')
 
     assert entity == fake_entity
@@ -46,10 +47,57 @@ async def test_should_get_one_from_cache(
         asynctest.CoroutineMock()
     )
     fake_service.cache['fake'] = fake_entity
+
     entity = await fake_service.get_one('fake', other_id='other_fake')
 
     assert entity == fake_entity
     assert not fake_service.repository.memory_data_source.hgetall.called
+
+
+@pytest.mark.asyncio
+async def test_should_get_one_from_fallback_when_not_found_on_memory(
+    fake_service, serialized_fake_entity, fake_entity
+):
+    client = fake_service.repository.fallback_data_source.client
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake'
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+    )
+
+    entity = await fake_service.get_one('fake', other_id='other_fake')
+
+    assert entity == fake_entity
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake'
+    )
+
+
+@pytest.mark.asyncio
+async def test_should_get_one_from_fallback_when_not_found_on_memory_with_fields(
+    fake_service, serialized_fake_entity, fake_entity
+):
+    client = fake_service.repository.fallback_data_source.client
+    await fake_service.repository.memory_data_source.delete(
+        'fake:other_fake:fake'
+    )
+    await fake_service.repository.fallback_data_source.put(
+        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+    )
+    fake_entity.number = None
+    fake_entity.boolean = False
+
+    entity = await fake_service.get_one(
+        'fake',
+        other_id='other_fake',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+    )
+
+    assert entity == fake_entity
+    assert fake_service.repository.memory_data_source.exists(
+        'fake:other_fake:fake'
+    )
 
 
 @pytest.mark.asyncio