+++ dbdaora/hash/repositories/_tests/test_integration_repository_hash_datastore.py
@@ -17,7 +17,7 @@ class FakeEntity:
 
 class FakeRepository(DatastoreHashRepository, entity_cls=FakeEntity):
     key_attrs = ('id',)
-    exclue_all_from_indexes = True
+    exclude_all_from_indexes = True
 
 
 @pytest.fixture
@@ -37,7 +37,7 @@ async def test_should_exclude_all_attributes_from_indexes(repository):
     entity = datastore.Entity(key=key)
     entity.update({'id': 'fake', 'fake_int': 1})
     client.put(entity)
-    time.sleep(0.5)
+    time.sleep(1)
     query = client.query(kind='fake')
     query.add_filter('fake_int', '=', 1)
 
@@ -46,7 +46,7 @@ async def test_should_exclude_all_attributes_from_indexes(repository):
 
     await repository.add(FakeEntity(id='fake', fake_int=1))
 
-    time.sleep(0.5)
+    time.sleep(1)
     query = client.query(kind='fake')
     query.add_filter('values', '=', 1)
     entities = query.fetch()