+++ dbdaora/hash/repositories/_tests/test_unit_repository_aioredis_hash_get_many_fields.py
@@ -17,13 +17,14 @@ async def test_should_set_already_not_found_error_when_get_many_with_fields(
         many=[fake_entity],
         fields=['id', 'integer', 'inner_entities'],
     )
-    pipeline = repository.memory_data_source.pipeline = mocker.MagicMock()
-    pipeline.return_value.execute = asynctest.CoroutineMock(
-        return_value=[None]
+    repository.memory_data_source.hmget = asynctest.CoroutineMock(
+        return_value=[None, None, None]
     )
-    pipeline.exists = asynctest.CoroutineMock(side_effect=[False])
-    repository.fallback_data_source.get_many = asynctest.CoroutineMock(
-        return_value=[None]
+    repository.memory_data_source.exists = asynctest.CoroutineMock(
+        side_effect=[False]
+    )
+    repository.fallback_data_source.get = asynctest.CoroutineMock(
+        return_value=None
     )
     repository.memory_data_source.set = asynctest.CoroutineMock()
     repository.memory_data_source.hmset = asynctest.CoroutineMock()
@@ -34,14 +35,14 @@ async def test_should_set_already_not_found_error_when_get_many_with_fields(
         ).entities
 
     assert exc_info.value.args == (expected_query,)
-    assert pipeline.return_value.hmget.call_args_list == [
+    assert repository.memory_data_source.hmget.call_args_list == [
         mocker.call('fake:fake', 'id', 'integer', 'inner_entities'),
     ]
-    assert pipeline.return_value.exists.call_args_list == [
+    assert repository.memory_data_source.exists.call_args_list == [
         mocker.call('fake:not-found:fake')
     ]
-    assert repository.fallback_data_source.get_many.call_args_list == [
-        mocker.call(['fake:fake'])
+    assert repository.fallback_data_source.get.call_args_list == [
+        mocker.call('fake:fake')
     ]
     assert repository.memory_data_source.set.call_args_list == [
         mocker.call('fake:not-found:fake', '1')